requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.3.3
cycler==0.12.1
ensemble-boxes==1.0.9
filelock==3.13.1
fonttools==4.60.1
fsspec==2024.6.1
idna==3.4
Jinja2==3.1.4
kiwisolver==1.4.9
llvmlite==0.45.1
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.3
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.0.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.28.1
scipy==1.16.2
six==1.17.0
sympy==1.13.3
torch==2.2.0+cu121
torchvision==0.17.0+cu121
triton==2.2.0
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.206
ultralytics-thop==2.0.17
urllib3==1.26.13
